requirements.txt
aiohappyeyeballs==2.6.1
aiohttp==3.12.6
aiosignal==1.3.2
altgraph==0.17.4
anyio==4.9.0
argcomplete==1.10.3
attrs==25.3.0
auto-py-to-exe==2.46.0
beautifulsoup4==4.8.2
blinker==1.9.0
bottle==0.13.2
bottle-websocket==0.2.9
branca==0.8.1
cefpython3==66.1
certifi==2025.1.31
cffi==1.17.1
chardet==3.0.4
charset-normalizer==3.4.1
click==8.1.8
clr_loader==0.2.7.post0
colorama==0.4.6
compressed-rtf==1.0.7
comtypes==1.4.11
docx2txt==0.9
ebcdic==1.1.1
Eel==0.18.1
extract-msg==0.28.7
Flask==3.1.0
folium==0.19.5
frozenlist==1.6.0
future==1.0.0
gevent==24.11.1
gevent-websocket==0.10.1
greenlet==3.1.1
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
idna==3.10
IMAPClient==2.1.0
itsdangerous==2.2.0
Jinja2==3.1.6
lxml==5.4.0
MarkupSafe==3.0.2
mgrs==1.5.0
MouseInfo==0.1.3
mss==10.0.0
multidict==6.4.4
numpy==2.2.5
olefile==0.47
opencv-python==4.11.0.86
outcome==1.3.0.post0
packaging==24.2
pdfminer.six==20191110
pefile==2023.2.7
pillow==10.4.0
playsound==1.3.0
plyer==2.1.0
propcache==0.3.1
proxy_tools==0.1.0
PyAutoGUI==0.9.54
pycparser==2.22
pycryptodome==3.23.0
PyGetWindow==0.0.9
pyinstaller==6.13.0
pyinstaller-hooks-contrib==2025.2
PyMsgBox==1.0.9
pyparsing==3.2.3
pyperclip==1.9.0
pypiwin32==223
pyproj==3.7.1
PyRect==0.2.0
PyScreeze==1.0.1
PySocks==1.7.1
pystray==0.19.5
python-docx==1.1.2
python-dotenv==1.1.0
python-pptx==0.6.23
python-telegram-bot==22.0
pythonnet==3.0.5
pytweening==1.2.0
pywebview==5.4
pywhatkit==5.4
pywin32==310
pywin32-ctypes==0.2.3
pywinauto==0.6.9
qrcode==8.2
requests==2.32.3
selenium==4.32.0
setuptools==78.1.0
six==1.12.0
sniffio==1.3.1
sortedcontainers==2.4.0
soupsieve==2.7
SpeechRecognition==3.8.1
textract==1.6.5
tkhtmlview==0.3.1
trio==0.30.0
trio-websocket==0.12.2
typing_extensions==4.13.1
tzdata==2025.2
tzlocal==5.3.1
ukrainealarm==0.0.1
urllib3==2.3.0
webdriver-manager==4.0.2
websocket-client==1.8.0
Werkzeug==3.1.3
wheel==0.45.1
wikipedia==1.4.0
win10toast==0.9
winotify==1.1.0
wsproto==1.2.0
xlrd==1.2.0
XlsxWriter==3.2.3
xyzservices==2025.4.0
yarl==1.20.0
zope.event==5.0
zope.interface==7.2
